refactor: field added tracking
